rename sections to sections-boxes
