+++ packages/react-layout/src/website/demos/index.js
@@ -6,13 +6,13 @@
 import itemLayouts from './item-layouts';
 import itemWizard from './item-wizard';
 import itemResponsive from './item-responsive';
-import sections from './sections';
+import sectionsBoxes from './sections-boxes';
 import sectionsGrid from './sections-grid';
 
 export default {
   itemLayouts,
   itemWizard,
   itemResponsive,
-  sections,
+  sectionsBoxes,
   sectionsGrid,
 };
