+++ packages/react-layout/src/website/demos/sections-boxes/demo.description.jsx
