+++ packages/react-layout/src/website/menu.js
@@ -24,7 +24,7 @@ export default [{
     getSubItem('item', 'layouts', 'item-layouts', 'Layouts', [docs.overview, docs.item]),
     getSubItem('item', 'responsive', 'item-responsive', 'Responsive', [docs.overview, docs.item]),
     getSubItem('item', 'wizard', 'item-wizard', 'Wizard', [docs.overview]),
-    getSubItem('item', 'sections', 'sections', 'Sections - Boxes', [docs.overview, docs.section]),
+    getSubItem('item', 'sections-boxes', 'sections-boxes', 'Sections - Boxes', [docs.overview, docs.section]),
     getSubItem('item', 'sections-grid', 'sections-grid', 'Sections - Grid', [docs.overview, docs.section]),
   ],
 }];
